Argument enabled python scripts
Added the ability to call Pull Data From Pdf with a system argument to a single file. Output will be stored in data folder.

Pull Data From Pdf can also be run without arguments. It will look in the IR samples folder for all pdf's and the query folder for a file named Query.pdf

Compare To Query.py now outputs a sorted list of closest matches using the print command. Items are separated by whitespace
